pyproject.toml
[tool.poetry]
name = "taleweaver"
version = "0.1.0"
description = ""
authors = ["Anand Badrinath"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
questionary = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
mypy = "^1.10.0"
ruff = "^0.4.4"
types-pyyaml = "^6.0.12.20240311"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.extend-select = ["I"]

[tool.pytest.ini_options]
pythonpath = ["src"]
